cofly/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pyjwt>=2.8.0
bcrypt>=4.0.0
protobuf>=4.25.0
websockets>=12.0

# test
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
requests>=2.32.5
